backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf==3.17.1
google-generativeai==0.3.1
python-dotenv==1.0.0
sqlalchemy==1.4.50
pydantic>=2.0.0 